Moved NavBar to components folder
