+++ frontend/src/components/NavBar.jsx
